Rename communication service to chat service
