+++ chat-service/.dockerignore
