+++ chat-service/Dockerfile
