+++ chat-service/package.json
