+++ docker-compose.yaml
@@ -69,10 +69,10 @@ services:
     depends_on:
       - mongo
 
-  communication-service:
-    container_name: communication-service
+  chat-service:
+    container_name: chat-service
     build:
-      context: ./communication-service
+      context: ./chat-service
     command: npm start
     ports:
       - "9000:9000"
@@ -110,7 +110,7 @@ services:
       - matching-service
       - user-service 
       - collab-service
-      - communication-service
+      - chat-service
       - video-service
